For unbounded memory areas, limit the size of the ByteBuffer to Integer.MAX_VALUE (2G)

+++ src/org/jruby/ext/ffi/jffi/NativeMemoryIO.java
@@ -28,7 +28,7 @@ class NativeMemoryIO implements MemoryIO, DirectMemoryIO {
     }
 
     public final java.nio.ByteBuffer asByteBuffer() {
-        return IO.newDirectByteBuffer(address, Long.MAX_VALUE);
+        return IO.newDirectByteBuffer(address, Integer.MAX_VALUE);
     }
 
     @Override
